@@ -2,20 +2,20 @@ name: navigation_manager
 description: >- 
   Navigation manager for Navigator 2.0 with a convenient declarative way 
   to define application routes.
-version: 0.1.0
+version: 1.0.0
 homepage: https://github.com/nogipx/navigation_manager
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   
-  uri: ^0.11.4
-  meta: ^1.3.0
-  equatable: ^1.0.0
+  uri: ^1.0.0
+  meta: any
+  equatable: ^2.0.3
 
 dev_dependencies:
   flutter_test:
